chore(ci): brancher 3 garde-fous soft KPI cycle 21 (CCN, coverage domain, ports migration)
